pyproject.toml
[tool.poetry]
name = "utaoi-radio"
version = "0.1.1"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
gspread = "^5.12.2"
oauth2client = "^4.1.3"
python-vlc = "^3.0.20123"
pafy = "^0.5.5"
youtube-dl = "^2021.12.17"
yt-dlp = "^2023.11.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
